requirements.txt
apscheduler==3.10.4
pyyaml==6.0
yfinance>=0.2.50
pandas==2.1.3
numpy==1.26.2
google-generativeai==0.3.0
requests
pykalman>=0.9.5
hmmlearn>=0.2.8
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
scipy>=1.11.0
